Raised Files to Root, Disabled SSR/Prerender, Fixed Status Filter
Project files moved from ./CodingTest to ./ to ensure proper project presentation.

SSR/Prerender attempts to set up a side server prior to running the backend and throws errors due to this. Since this is a locally run static spa there's really no need to keep doing this.

A conflict in naming conventions of each status type between the backend and frontend causes the filter to incorrectly look for or sort through the tasks. Naming convention changed to correctly associated the same variable between both ends.